Refactor out common validation code
fixes #47
